<refact>: use fastapi router
